refactor: Renewal structure for User-Service and Auth-Service
