+++ package.json
@@ -11,13 +11,13 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@prisma/client": "^6.15.0",
-    "express": "^5.1.0",
-    "prisma": "^6.15.0"
+    "express": "^5.1.0"
   },
   "devDependencies": {
+    "@prisma/client": "^6.16.1",
     "@types/express": "^5.0.3",
     "@types/node": "^24.3.0",
+    "prisma": "^6.16.1",
     "ts-node-dev": "^2.0.0",
     "typescript": "^5.9.2"
   }
