+++ services/auth-service/.gitignore
@@ -1,0 +1,49 @@
+# Node modules
+node_modules/
+
+# Build output
+dist/
+build/
+
+# TypeScript cache
+*.tsbuildinfo
+
+# Logs
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+pnpm-debug.log*
+logs/
+*.log
+
+# Environment variables
+.env
+.env.*.local
+
+# OS files
+.DS_Store
+Thumbs.db
+
+# IDE files
+.vscode/
+.idea/
+*.swp
+
+# Coverage & testing
+coverage/
+.nyc_output/
+
+# Temporary
+.tmp/
+.cache/
+
+# Optional lock files (tùy dự án có thể giữ lại 1)
+package-lock.json
+yarn.lock
+pnpm-lock.yaml
+
+/gateway/node_modules
+
+/src/generated/prisma
+
+/src/generated/prisma
