+++ services/auth-service/package.json
@@ -1,32 +1,43 @@
 {
-  "name": "user-sevice",
+  "name": "auth-service",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon src/index.ts"
+    "dev": "nodemon src/index.ts",
+    "seed": "ts-node prisma/seed.ts"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@types/express": "^5.0.3",
-    "@types/express-validator": "^2.20.33",
+    "@prisma/client": "^6.15.0",
     "@types/helmet": "^0.0.48",
     "@types/jsonwebtoken": "^9.0.10",
     "@types/lodash": "^4.17.20",
     "@types/morgan": "^1.9.10",
-    "@types/node": "^24.3.3",
+    "@types/nodemailer": "^7.0.1",
     "cloudinary": "^2.7.0",
-    "dotenv": "^17.2.2",
+    "crypto": "^1.0.1",
     "express": "^5.1.0",
     "express-validator": "^7.2.1",
     "helmet": "^8.1.0",
     "jsonwebtoken": "^9.0.2",
     "lodash": "^4.17.21",
     "morgan": "^1.10.1",
+    "nodemailer": "^7.0.6",
+    "prisma": "^6.15.0"
+  },
+  "devDependencies": {
+    "@types/bcrypt": "^6.0.0",
+    "@types/express": "^5.0.3",
+    "@types/express-validator": "^2.20.33",
+    "@types/node": "^24.3.0",
+    "bcrypt": "^6.0.0",
+    "dotenv": "^17.2.1",
     "nodemon": "^3.1.10",
-    "ts-node": "^10.9.2"
+    "ts-node": "^10.9.2",
+    "typescript": "^5.9.2"
   }
 }
