+++ services/auth-service/src/helpers/asyncHandler.ts
@@ -1,0 +1,7 @@
+import { Request, Response, NextFunction } from 'express'
+
+export const asyncHandler = (fn : any) => {
+    return (req: Request, res: Response, next: NextFunction) => {
+        fn(req, res, next).catch(next)
+    }
+}
