+++ services/auth-service/src/index.ts
@@ -1,27 +1,31 @@
-import express from 'express';
+import express, { Request, Response } from 'express';
 import { prisma } from '../../../config/database.config';
-import userRouter from './routes/user.routes'
+import authRouter from './routes/auth.routes'
 import helmet from 'helmet';
 import morgan from 'morgan';
+import { validateRefreshTokenMiddleware } from './utils/authToken';
 
 import dotenv from 'dotenv';
+import { authAccessTokenMiddleware } from './middleware/authMiddleware';
+
 dotenv.config();
 
+
 const app = express()
-const PORT = process.env.PORT || 3002
+const PORT = process.env.PORT || 3001
 
 app.use(helmet());
 app.use(morgan('combined'));
 app.use(express.json())
 app.use((req, res, next) => {
-    console.log(`User-service received: ${req.method} ${req.url}`);
-    next();
-})
+  console.log(`Auth-service received: ${req.method} ${req.url}`);
+  next();
+});
 
-app.use('/', userRouter) 
+app.use('/', authRouter)
 
-app.get('/test-server', (req, res) => {
-    res.send('User Service is running')
+app.get('/test-server', authAccessTokenMiddleware, (req, res) => {
+  res.send('Auth Service is running')
 })
 
 app.listen(PORT, () => {
