+++ services/auth-service/src/middleware/authMiddleware.ts
