+++ services/auth-service/src/middleware/errorHandler.ts
@@ -1,0 +1,10 @@
+import { Request, Response, NextFunction } from "express";
+
+export const errorHandler = (err: any, req: Request, res: Response, next: NextFunction) => {
+    console.error('Error caught: ', err.message)
+
+    res.status(err.status || 500).json({
+        success: false,
+        message: err.message || 'Internal Server Error',
+    })
+}
