+++ services/auth-service/src/models/repository/auth.repository.ts
@@ -1,4 +1,4 @@
-import { prisma } from '../../config/database.config'
+import { prisma } from '../../../../../config/database.config'
 
 class UserRepository {
