+++ services/auth-service/src/types/express.d.ts
@@ -1,0 +1,16 @@
+// Extend Request interface để TypeScript hiểu
+declare global {
+    namespace Express {
+        interface Request {
+            user?: {
+                id: number;
+                email: string;
+                name: string | null;
+                role: string;
+            };
+            token?: string;
+        }
+    }
+}
+
+export {}
