+++ services/auth-service/src/types/interface.ts
@@ -1,0 +1,7 @@
+// Interface cho JWT payload
+export interface JWTPayload {
+    userId: string;
+    email: string;
+    iat: number;
+    exp: number;
+}
