+++ services/auth-service/src/utils/authToken.ts
@@ -1,6 +1,6 @@
 import { Request, Response, NextFunction } from "express";
 import crypto from "crypto";
-import { prisma } from "../config/database.config";
+import { prisma } from "../../../../config/database.config";
 
 const hashToken = (token: string) =>
   crypto.createHash("sha256").update(token).digest("hex");
