+++ services/auth-service/src/utils/authUtils.ts
