+++ services/auth-service/src/utils/getInfoData.ts
@@ -1,0 +1,5 @@
+import _ from 'lodash';
+
+export const getInfoData = (fields: string[], data: any) => {
+    return _.pick(data, fields)
+}
