+++ services/auth-service/src/utils/hashPassword.ts
