+++ services/auth-service/src/utils/logger.ts
