+++ services/auth-service/src/utils/token.ts
