+++ services/auth-service/tsconfig.json
@@ -1,0 +1,35 @@
+{
+  "compilerOptions": {
+    "module": "commonjs",
+    "target": "ES6",
+    "types": ["node"],
+    "sourceMap": true,
+    "declaration": true,
+    "declarationMap": true,
+    "noUncheckedIndexedAccess": true,
+    "exactOptionalPropertyTypes": true,
+    "strict": true,
+    "jsx": "react-jsx",
+    "isolatedModules": true,
+    "noUncheckedSideEffectImports": true,
+    "moduleDetection": "force",
+    "moduleResolution": "node",
+    "skipLibCheck": true,
+    "esModuleInterop": true,
+    "forceConsistentCasingInFileNames": true,
+    "rootDir": "../../",
+    "outDir": "./dist",
+    "baseUrl": "../../",
+    "paths": {
+      "@shared/*": ["shared/*"]
+    }
+  },
+  "include": [
+    "src/**/*",
+    "../../shared/**/*"
+, "../../config/database.config.ts", "../user-service/src/config/cloudinary.config.ts", "../user-service/src/models/profileUserModel.ts"  ],
+  "exclude": [
+    "node_modules",
+    "dist"
+  ]
+}
