+++ services/user-service/.gitignore
@@ -41,9 +41,3 @@ coverage/
 package-lock.json
 yarn.lock
 pnpm-lock.yaml
-
-/gateway/node_modules
-
-/src/generated/prisma
-
-/src/generated/prisma
