+++ services/user-service/src/routes/user.routes.ts
@@ -1,24 +1,12 @@
 import UserController from "../controllers/user.controller";
-import express from 'express'
-import { check, checkSchema } from "express-validator";
-import { validateSignup, validateLogin } from "../middleware/validationInput";
-import { validateRefreshTokenMiddleware } from "../utils/authToken";
-import { asyncHandler } from '../helpers/asyncHandler'
-import { errorHandler } from "../middleware/errorHandler";
-import { resetTokenMiddleware } from "../middleware/authResetTokenMiddleware";
+import express from 'express';
+import { asyncHandler } from "../helpers/asyncHandler";
 import { authAccessTokenMiddleware } from "../middleware/authMiddleware";
+import { errorHandler } from "../middleware/errorHandler";
 
 const router = express.Router()
 
-router.post('/signup', checkSchema(validateSignup), asyncHandler(UserController.register))
-router.post('/login', checkSchema(validateLogin), asyncHandler(UserController.login))
-router.post('/logout', validateRefreshTokenMiddleware, asyncHandler(UserController.logout))
-
-router.post('/forgot-password', asyncHandler(UserController.forgotPassword))
-router.post('/reset-password', resetTokenMiddleware, asyncHandler(UserController.resetPassword))
-
 router.patch('/update-profile', authAccessTokenMiddleware, asyncHandler(UserController.updateProfile))
-router.post('/refresh-accessToken', validateRefreshTokenMiddleware, asyncHandler(UserController.refreshAccessToken))
 
 router.use(errorHandler)
