+++ services/user-service/src/types/interface.ts
